Refactor: ts마이그레이션(user서비스)
